pkg/endpoint: move base64-related functions to base64.go
Signed-off by: Ian Vernon <ian@cilium.io>
