+++ pkg/endpoint/base64.go
@@ -1,0 +1,77 @@
+// Copyright 2018 Authors of Cilium
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package endpoint
+
+import (
+	"encoding/base64"
+	"encoding/json"
+	"fmt"
+	"strings"
+)
+
+// base64 returns the endpoint in a base64 format.
+func (e *Endpoint) base64() (string, error) {
+	var (
+		jsonBytes []byte
+		err       error
+	)
+	if e.Consumable != nil {
+		e.Consumable.Mutex.RLock()
+		jsonBytes, err = json.Marshal(e)
+		e.Consumable.Mutex.RUnlock()
+	} else {
+		jsonBytes, err = json.Marshal(e)
+	}
+	if err != nil {
+		return "", err
+	}
+	return base64.StdEncoding.EncodeToString(jsonBytes), nil
+}
+
+// parseBase64ToEndpoint parses the endpoint stored in the given base64 string.
+func parseBase64ToEndpoint(str string, ep *Endpoint) error {
+	jsonBytes, err := base64.StdEncoding.DecodeString(str)
+	if err != nil {
+		return err
+	}
+	return json.Unmarshal(jsonBytes, ep)
+}
+
+// ParseEndpoint parses the given strEp which is in the form of:
+// common.CiliumCHeaderPrefix + common.Version + ":" + endpointBase64
+func ParseEndpoint(strEp string) (*Endpoint, error) {
+	// TODO: Provide a better mechanism to update from old version once we bump
+	// TODO: cilium version.
+	strEpSlice := strings.Split(strEp, ":")
+	if len(strEpSlice) != 2 {
+		return nil, fmt.Errorf("invalid format %q. Should contain a single ':'", strEp)
+	}
+	var ep Endpoint
+	if err := parseBase64ToEndpoint(strEpSlice[1], &ep); err != nil {
+		return nil, fmt.Errorf("failed to parse base64toendpoint: %s", err)
+	}
+
+	// We need to check for nil in Status, CurrentStatuses and Log, since in
+	// some use cases, status will be not nil and Cilium will eventually
+	// error/panic if CurrentStatus or Log are not initialized correctly.
+	// Reference issue GH-2477
+	if ep.Status == nil || ep.Status.CurrentStatuses == nil || ep.Status.Log == nil {
+		ep.Status = NewEndpointStatus()
+	}
+
+	ep.state = StateRestoring
+
+	return &ep, nil
+}
